Image_In_Terminal/requirements.txt
color-it
pillow